Remove unnecessary unboxing.
Explicit unboxing has been unnecessary since Java 5, so this is safely removable.

+++ main/src/paulscode/android/mupen64plus/MenuSettingsCoreConfigureActivity.java
@@ -52,7 +52,7 @@ public class MenuSettingsCoreConfigureActivity extends ListActivity implements I
         {
             try
             {  // Make sure a valid integer was entered
-                scancode = Integer.valueOf( val ).intValue();
+                scancode = Integer.valueOf( val );
             }
             catch( NumberFormatException nfe )
             {}  // Skip it if this happens
